properly handle file conflicts for non-merging games
currently this affects Pillars of Eternity 2 and Kingdom Come: Deliverance.
In PoE2 mods aren't merged and each mod is deployed to a directory, named identically to the mod id inside vortex - which is guaranteed to be unique.
Therefore file conflicts aren't possible and won't be reported.
In KC:D mods are deployed to a directories _derived_ from the mod id but because of that it's possible multiple mods do get merged into the same subdirectory and then they can conflict.
This extension will now correctly report only actual file conflicts caused by the deploy directory being identical.