detect file conflicts across mod types
in practice previously we might have detected file conflicts that weren't conflicts if two mods with different mod types contained the same file but due to their mod type it was deployed to different directories, which almost always meant that they _should_ conflict but one of the mod types was incorrect.

the conflict check now more correctly reflects actual file conflicts, for that all files are compared based on their final output directory. This also means that when presenting file conflicts we display the absolute path