fixed conflict rules not being maintained when reinstalling/updating mod
(this bug was never released to the public)

Previously we had a very basic mod remove event listener in charge of
removing conflict rules referencing a mod that gets removed by the user.
The problem was, the event listener would execute when reinstalling or
updating a mod which would remove the rules. Obviously that wasn't
intentional, the rules should only get removed if the mod is removed
entirely.