improved grouped rule assignment to update the local state sparingly
Previously using before/after all to assign a rule to multiple mods when
viewing the conflicts of a single mod would execute repeated calls to
the nextState proxy in order to update the component - that could
potentially cause some updates to be missed. We now update the component
rules state property only once.