improved how we handle orphaned conflict removal on update/reinstall
We no longer use the notifications state branch to check if a mod is
being replaced during reinstall/update.

A new 'options' parameter has been added to the remove-mod event which
can be used to pass on relevant data to event listeners. It's currently
very basic but could be extended in the future.
{
  willBeReplaced: boolean;
}