fixed: filtering by dependencies broken if column disabled
if clicking the icon left of a mod with dependencies it would enable a filter for that mods dependencies and start displaying the "indent" icon alongside every visible mod.
But if the dependencies column wasn't actually enabled the filter didn't apply so the icon was displayed on all mods.
Now the icon won't show if the column is disabled and it does get enabled when clicking the icon.

fixes Nexus-Mods/Vortex#11825